+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/experience/ExperienceActivity.kt
@@ -10,7 +10,6 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.text.TextUtils
 import android.view.KeyEvent
 import android.view.View
 import android.view.ViewGroup
@@ -198,7 +197,7 @@ open class ExperienceActivity : BaseExperienceActivity(), StartReactInstanceDele
             emitUpdatesEvent(params)
           }
 
-          override fun updateStatus(status: AppLoaderStatus) {
+          override fun updateStatus(status: AppLoaderStatus?) {
             setLoadingProgressStatusIfEnabled(status)
           }
 
@@ -442,7 +441,7 @@ open class ExperienceActivity : BaseExperienceActivity(), StartReactInstanceDele
 
     // Sometime we want to release a new version without adding a new .aar. Use TEMPORARY_SDK_VERSION
     // to point to the unversioned code in ReactAndroid.
-    if (Constants.TEMPORARY_SDK_VERSION == sdkVersion) {
+    if (Constants.SDK_VERSION == sdkVersion) {
       sdkVersion = RNObject.UNVERSIONED
     }
 
@@ -450,17 +449,10 @@ open class ExperienceActivity : BaseExperienceActivity(), StartReactInstanceDele
     detachSdkVersion = sdkVersion
 
     if (RNObject.UNVERSIONED != sdkVersion) {
-      var isValidVersion = false
-      for (version in Constants.SDK_VERSIONS_LIST) {
-        if (version == sdkVersion) {
-          isValidVersion = true
-          break
-        }
-      }
+      val isValidVersion = sdkVersion == Constants.SDK_VERSION
       if (!isValidVersion) {
         KernelProvider.instance.handleError(
-          sdkVersion + " is not a valid SDK version. Options are " +
-            TextUtils.join(", ", Constants.SDK_VERSIONS_LIST) + ", " + RNObject.UNVERSIONED + "."
+          sdkVersion + " is not a valid SDK version. Only ${Constants.SDK_VERSION} is supported."
         )
         return
       }