+++ packages/expo-notifications/src/Notifications.types.ts
@@ -165,6 +165,18 @@ export interface WeeklyNotificationTrigger {
 }
 
 /**
+ * A trigger related to a monthly notification.
+ * > The same functionality will be achieved on iOS with a `CalendarNotificationTrigger`.
+ * @platform android
+ */
+export interface MonthlyNotificationTrigger {
+  type: 'monthly';
+  day: number;
+  hour: number;
+  minute: number;
+}
+
+/**
  * A trigger related to a yearly notification.
  * > The same functionality will be achieved on iOS with a `CalendarNotificationTrigger`.
  * @platform android
@@ -242,6 +254,7 @@ export type NotificationTrigger =
   | TimeIntervalNotificationTrigger
   | DailyNotificationTrigger
   | WeeklyNotificationTrigger
+  | MonthlyNotificationTrigger
   | YearlyNotificationTrigger
   | UnknownNotificationTrigger;
 
@@ -260,6 +273,7 @@ export enum SchedulableTriggerInputTypes {
   CALENDAR = 'calendar',
   DAILY = 'daily',
   WEEKLY = 'weekly',
+  MONTHLY = 'monthly',
   YEARLY = 'yearly',
   DATE = 'date',
   TIME_INTERVAL = 'timeInterval',
@@ -311,6 +325,19 @@ export type WeeklyTriggerInput = {
   type: SchedulableTriggerInputTypes.WEEKLY;
   channelId?: string;
   weekday: number;
+  hour: number;
+  minute: number;
+};
+
+/**
+ * This trigger input will cause the notification to be delivered once per month
+ * when the `day`, `hour`, and `minute` date components match the specified values.
+ * > **Note:** All properties are specified in JavaScript `Date` object's ranges.
+ */
+export type MonthlyTriggerInput = {
+  type: SchedulableTriggerInputTypes.MONTHLY;
+  channelId?: string;
+  day: number;
   hour: number;
   minute: number;
 };
@@ -369,6 +396,7 @@ export type SchedulableNotificationTriggerInput =
   | TimeIntervalTriggerInput
   | DailyTriggerInput
   | WeeklyTriggerInput
+  | MonthlyTriggerInput
   | YearlyTriggerInput
   | DateTriggerInput;