+++ packages/expo-video/src/index.ts
@@ -14,4 +14,5 @@ export {
   VideoMetadata,
   DRMType,
   DRMOptions,
+  BufferOptions,
 } from './VideoPlayer.types';